requirements.txt
h5py==3.10.0
numpy==1.24.1
opencv-python==4.9.0.80
pynput
pillow==10.2.0
pyyaml
stable-baselines3==2.2.1
tensorboard
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
transformers==4.38.2